Move 'CardBody' subcomponent to `card-body` subfolder, refactor hook out
