+++ packages/components/src/card/card-body/README.md
@@ -1,0 +1,7 @@
+# CardBody
+
+<div class="callout callout-alert">
+This feature is still experimental. “Experimental” means this is an early implementation subject to drastic and breaking changes.
+</div>
+
+TBD
