+++ packages/components/src/card/card-body/index.js
@@ -1,0 +1,2 @@
+export { default } from './component';
+export { useCardBody } from './hook';
