+++ packages/components/src/card/index.js
@@ -1,5 +1,4 @@
 export { default as Card, useCard } from './card';
 export { default as CardHeader, useCardHeader } from './card-header';
 export { default as CardFooter, useCardFooter } from './card-footer';
-
-export { default as CardBody } from './body';
+export { default as CardBody, useCardBody } from './card-body';
